rename to make it explicit that it's related to config
