+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/config/MapperConfig.java
@@ -1,8 +1,8 @@
 package br.com.jpcchaves.infrastructure.config;
 
 import br.com.jpcchaves.infrastructure.mapper.CategoryEntityMapper;
-import br.com.jpcchaves.infrastructure.mapper.TodoRequestDTOMapper;
-import br.com.jpcchaves.infrastructure.mapper.TodoResponseDTOMapper;
+import br.com.jpcchaves.infrastructure.mapper.config.TodoRequestDTOConfigMapper;
+import br.com.jpcchaves.infrastructure.mapper.config.TodoResponseDTOConfigMapper;
 import org.modelmapper.ModelMapper;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
@@ -16,8 +16,8 @@ public class MapperConfig {
   public ModelMapper modelMapper() {
     ModelMapper modelMapper = new ModelMapper();
 
-    modelMapper.addMappings(new TodoRequestDTOMapper());
-    modelMapper.addMappings(new TodoResponseDTOMapper());
+    modelMapper.addMappings(new TodoRequestDTOConfigMapper());
+    modelMapper.addMappings(new TodoResponseDTOConfigMapper());
     modelMapper.addMappings(new CategoryEntityMapper());
 
     return modelMapper;
