+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/mapper/config/TodoRequestDTOConfigMapper.java
@@ -1,10 +1,11 @@
-package br.com.jpcchaves.infrastructure.mapper;
+package br.com.jpcchaves.infrastructure.mapper.config;
 
 import br.com.jpcchaves.core.domain.Todo;
 import br.com.jpcchaves.infrastructure.dto.todo.TodoRequestDTO;
 import org.modelmapper.PropertyMap;
 
-public class TodoRequestDTOMapper extends PropertyMap<TodoRequestDTO, Todo> {
+public class TodoRequestDTOConfigMapper extends PropertyMap<TodoRequestDTO, Todo> {
+
   protected void configure() {
     skip().setId(null);
   }
