+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/mapper/config/TodoResponseDTOConfigMapper.java
@@ -1,10 +1,11 @@
-package br.com.jpcchaves.infrastructure.mapper;
+package br.com.jpcchaves.infrastructure.mapper.config;
 
 import br.com.jpcchaves.core.domain.Todo;
 import br.com.jpcchaves.infrastructure.dto.todo.TodoResponseDTO;
 import org.modelmapper.PropertyMap;
 
-public class TodoResponseDTOMapper extends PropertyMap<Todo, TodoResponseDTO> {
+public class TodoResponseDTOConfigMapper extends PropertyMap<Todo, TodoResponseDTO> {
+
   @Override
   protected void configure() {
     map().setCategory(source.getCategory().getName());
